requirements.txt
torch>=1.12.0
torchvision>=0.13.0
accelerate>=0.20.0
swanlab>=0.3.0
pyyaml>=6.0
tqdm>=4.64.0
numpy>=1.21.0
timm>=0.9.0
torchinfo>=1.8.0
